new diagram representation
